+++ dartsim/src/SDFFeatures_TEST.cc
@@ -41,6 +41,8 @@
 #include <sdf/Root.hh>
 #include <sdf/World.hh>
 
+#include <test/Utils.hh>
+
 struct TestFeatureList : ignition::physics::FeatureList<
     ignition::physics::GetBasicJointState,
     ignition::physics::SetBasicJointState,
@@ -314,6 +316,186 @@ TEST(SDFFeatures_TEST, FallbackToFixedJoint)
         dynamic_cast<const dart::dynamics::WeldJoint *>(joint);
     EXPECT_NE(nullptr, fixedJoint) << " joint type is: " << joint->getType();
   }
+}
+
+TEST(SDFFeatures_FrameSemantics, LinkRelativeTo)
+{
+  World world = LoadWorld(TEST_WORLD_DIR"/model_frames.sdf");
+  const std::string modelName = "link_relative_to";
+
+  dart::simulation::WorldPtr dartWorld = world.GetDartsimWorld();
+  ASSERT_NE(nullptr, dartWorld);
+
+  const dart::dynamics::SkeletonPtr skeleton =
+      dartWorld->getSkeleton(modelName);
+
+  ASSERT_NE(nullptr, skeleton);
+  ASSERT_EQ(2u, skeleton->getNumBodyNodes());
+
+
+  const dart::dynamics::BodyNode *link2 = skeleton->getBodyNode("L2");
+  ASSERT_NE(nullptr, link2);
+
+  // Expect the world pose of L2 to be 0 0 3 0 0 pi
+  Eigen::Isometry3d expWorldPose =
+      Eigen::Translation3d(0, 0, 3) *
+      Eigen::AngleAxisd(IGN_PI, Eigen::Vector3d::UnitZ());
+
+  dartWorld->step();
+
+  // Step once and check
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      expWorldPose, link2->getWorldTransform(), 1e-3));
+}
+
+TEST(SDFFeatures_FrameSemantics, CollisionRelativeTo)
+{
+  World world = LoadWorld(TEST_WORLD_DIR"/model_frames.sdf");
+  const std::string modelName = "collision_relative_to";
+
+  dart::simulation::WorldPtr dartWorld = world.GetDartsimWorld();
+  ASSERT_NE(nullptr, dartWorld);
+
+  const dart::dynamics::SkeletonPtr skeleton =
+      dartWorld->getSkeleton(modelName);
+
+  ASSERT_NE(nullptr, skeleton);
+  ASSERT_EQ(2u, skeleton->getNumBodyNodes());
+
+
+  const dart::dynamics::BodyNode *link2 = skeleton->getBodyNode("L2");
+  ASSERT_NE(nullptr, link2);
+
+  const auto collision = link2->getShapeNode(0);
+  ASSERT_TRUE(collision);
+  // Expect the pose of c1 relative to L2 (the parent link) to be the same
+  // as the pose of L1 relative to L2
+  Eigen::Isometry3d expPose;
+  expPose = Eigen::Translation3d(0, 0, -1);
+
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      expPose, collision->getRelativeTransform(), 1e-5));
+
+  // Step once and check, the relative pose should still be the same
+  dartWorld->step();
+
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      expPose, collision->getRelativeTransform(), 1e-5));
+}
+
+TEST(SDFFeatures_FrameSemantics, ExplicitFramesWithLinks)
+{
+  World world = LoadWorld(TEST_WORLD_DIR"/model_frames.sdf");
+  const std::string modelName = "explicit_frames_with_links";
+
+  dart::simulation::WorldPtr dartWorld = world.GetDartsimWorld();
+  ASSERT_NE(nullptr, dartWorld);
+
+  const dart::dynamics::SkeletonPtr skeleton =
+      dartWorld->getSkeleton(modelName);
+
+  ASSERT_NE(nullptr, skeleton);
+  ASSERT_EQ(2u, skeleton->getNumBodyNodes());
+
+
+  const dart::dynamics::BodyNode *link1 = skeleton->getBodyNode("L1");
+  ASSERT_NE(nullptr, link1);
+
+  const dart::dynamics::BodyNode *link2 = skeleton->getBodyNode("L2");
+  ASSERT_NE(nullptr, link2);
+
+  // Expect the world pose of L1 to be the same as the world pose of F1
+  Eigen::Isometry3d link1ExpPose;
+  link1ExpPose = Eigen::Translation3d(1, 0, 1);
+
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      link1ExpPose, link1->getWorldTransform(), 1e-5));
+
+  // Expect the world pose of L2 to be the same as the world pose of F2, which
+  // is at the origin of the model
+  Eigen::Isometry3d link2ExpPose;
+  link2ExpPose = Eigen::Translation3d(1, 0, 0);
+
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      link2ExpPose, link2->getWorldTransform(), 1e-5));
+
+  // Step once and check
+  dartWorld->step();
+
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      link1ExpPose, link1->getWorldTransform(), 1e-5));
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      link2ExpPose, link2->getWorldTransform(), 1e-5));
+}
+
+TEST(SDFFeatures_FrameSemantics, ExplicitFramesWithCollision)
+{
+  World world = LoadWorld(TEST_WORLD_DIR"/model_frames.sdf");
+  const std::string modelName = "explicit_frames_with_collisions";
+
+  dart::simulation::WorldPtr dartWorld = world.GetDartsimWorld();
+  ASSERT_NE(nullptr, dartWorld);
+
+  const dart::dynamics::SkeletonPtr skeleton =
+      dartWorld->getSkeleton(modelName);
+
+  ASSERT_NE(nullptr, skeleton);
+  ASSERT_EQ(1u, skeleton->getNumBodyNodes());
+
+  const dart::dynamics::BodyNode *link1 = skeleton->getBodyNode("L1");
+  ASSERT_NE(nullptr, link1);
+
+  const auto collision = link1->getShapeNode(0);
+  ASSERT_TRUE(collision);
+
+  // Expect the pose of c1 relative to L1 (the parent link) to be the same
+  // as the pose of F1 relative to L1
+  Eigen::Isometry3d expPose;
+  expPose = Eigen::Translation3d(0, 0, 1);
+
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      expPose, collision->getRelativeTransform(), 1e-5));
+
+  // Step once and check
+  dartWorld->step();
+
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      expPose, collision->getRelativeTransform(), 1e-5));
+}
+
+TEST(SDFFeatures_FrameSemantics, ExplicitWorldFrames)
+{
+  World world = LoadWorld(TEST_WORLD_DIR"/world_frames.sdf");
+  const std::string modelName = "M";
+
+  dart::simulation::WorldPtr dartWorld = world.GetDartsimWorld();
+  ASSERT_NE(nullptr, dartWorld);
+
+  const dart::dynamics::SkeletonPtr skeleton =
+      dartWorld->getSkeleton(modelName);
+
+
+  ASSERT_NE(nullptr, skeleton);
+  ASSERT_EQ(1u, skeleton->getNumBodyNodes());
+
+  const dart::dynamics::BodyNode *link1 = skeleton->getBodyNode("L1");
+  ASSERT_NE(nullptr, link1);
+
+  // Expect the world pose of M to be (1 1 2 0 0 0) taking into acount the chain
+  // of explicit frames relative to which its pose is expressed
+  Eigen::Isometry3d expPose;
+  expPose = Eigen::Translation3d(1, 1, 2);
+
+  // Since we can't get the skeleton's world transform, we use the world
+  // transform of L1 which is at the origin of the model frame.
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      expPose, link1->getWorldTransform(), 1e-5));
+
+  // Step once and check
+  dartWorld->step();
+
+  EXPECT_TRUE(ignition::physics::test::Equal(
+      expPose, link1->getWorldTransform(), 1e-5));
 }
 
 int main(int argc, char *argv[])