+++ Changelog.md
@@ -4,6 +4,9 @@
 
 ### Ignition Physics 2.0.0 (20XX-XX-XX)
 
+1. Add support for sdformat frame semantics
+    * [Pull request 106](https://bitbucket.org/ignitionrobotics/ign-physics/pull-requests/106)
+
 1. Upgrade to libsdformat9
     * [Pull request 108](https://bitbucket.org/ignitionrobotics/ign-physics/pull-requests/108)